backend/requirements.txt
setuptools>=65.0.0,<81
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt>=5.5.0
psycopg2-binary>=2.9.9
Pillow>=11.0.0
python-decouple==3.8
django-filter==23.5
celery==5.3.4
redis==5.0.1
requests==2.31.0
twilio==8.10.3
stripe==7.7.0
python-dotenv==1.0.0
# API Documentation
drf-yasg==1.21.7
# PDF Generation
reportlab==4.0.7
weasyprint==60.2
PyMuPDF==1.23.8
# File handling
django-storages==1.14.2
# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
# Validation
django-phonenumber-field==7.1.0
phonenumbers==8.13.26
# Date handling
python-dateutil==2.8.2